data transfer objects

+++ src/main/java/com/mopahta/projectmanager/service/ProjectTaskService.java
@@ -1,6 +1,7 @@
 package com.mopahta.projectmanager.service;
 
 import com.mopahta.projectmanager.dto.ProjectTaskDTO;
+import com.mopahta.projectmanager.exception.InvalidValuesException;
 import com.mopahta.projectmanager.exception.NotFoundException;
 import com.mopahta.projectmanager.model.ProjectTask;
 import com.mopahta.projectmanager.model.ProjectTaskKey;
@@ -8,6 +9,7 @@ import com.mopahta.projectmanager.repository.ProjectTaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -18,6 +20,20 @@ public class ProjectTaskService {
 
     @Autowired
     ProjectService projectService;
+
+    public List<ProjectTaskDTO> projectTasksToDTO(List<ProjectTask> projectTasks) {
+        List<ProjectTaskDTO> projectTasksDTO = new ArrayList<>();
+
+        projectTasks.forEach((ProjectTask projectTask) -> {
+            projectTasksDTO.add(
+                    new ProjectTaskDTO(
+                            projectTask.getId().getId(),
+                            projectTask.getId().getProjectId(),
+                            projectTask.getTask(),
+                            projectTask.isFinished()));
+        });
+        return projectTasksDTO;
+    }
 
     private ProjectTask DTOToProjectTask(ProjectTaskDTO projectTaskDTO) throws NotFoundException {
         ProjectTask task = new ProjectTask();
@@ -38,7 +54,15 @@ public class ProjectTaskService {
         projectTaskRepository.save(DTOToProjectTask(projectTaskDTO));
     }
 
-    public void removeTaskFromProject(ProjectTaskDTO projectTaskDTO) {
+    public void removeTaskFromProject(ProjectTaskDTO projectTaskDTO) throws InvalidValuesException {
+        if (!projectTaskRepository.existsById(new ProjectTaskKey(
+                projectTaskDTO.getOrderId(),
+                projectTaskDTO.getProjectId()
+        ))) {
+            throw new InvalidValuesException(
+                    "No task matched to project " + projectTaskDTO.getProjectId()
+                            + " order id " + projectTaskDTO.getOrderId());
+        }
         projectTaskRepository.deleteById(
                 new ProjectTaskKey(projectTaskDTO.getOrderId(), projectTaskDTO.getProjectId())
         );
